backend/requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
Brotli==1.1.0
certifi==2025.4.26
cffi==1.17.1
channels==4.2.2
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
constantly==23.10.4
contourpy==1.3.2
cryptography==44.0.3
cssselect2==0.8.0
cycler==0.12.1
daphne==4.1.2
defusedxml==0.7.1
distlib==0.3.9
distro==1.7.0
Django==5.2.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
filelock==3.18.0
fonttools==4.57.0
fpdf==1.7.2
fpdf2==2.8.3
greenlet==3.2.2
html5lib==1.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
kiwisolver==1.4.8
lxml==5.4.0
matplotlib==3.10.3
numpy==2.2.5
oscrypto==1.3.0
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pipenv==2025.0.2
platformdirs==4.3.8
playwright==1.52.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydyf==0.11.0
pyee==13.0.0
pyHanko==0.26.0
pyhanko-certvalidator==0.26.8
PyJWT==2.9.0
pyOpenSSL==25.0.0
pyparsing==3.2.3
pypdf==5.4.0
pyphen==0.17.2
python-bidi==0.6.6
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
qrcode==8.2
reportlab==4.4.0
requests==2.32.3
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.3
ssh-import-id==5.11
supervisor==4.2.1
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.4.0
virtualenv==20.31.2
weasyprint==65.1
webencodings==0.5.1
xhtml2pdf==0.2.17
zope.interface==7.2
zopfli==0.2.3.post1
